+++ lib/src/libc/string.c
