+++ lib/src/os/vitbl.c
