+++ tools/python/generate_ld.py
@@ -22,7 +22,7 @@ BUILD_DIR = 'build/' + VERSION
 
 LATE_DATA_FILES = [ 
     BUILD_DIR + '/lib/src/al/alCopy.o',
-    BUILD_DIR + '/lib/src/printf.o',
+    BUILD_DIR + '/lib/src/libc/xprintf.o',
     BUILD_DIR + '/lib/src/os/osTimer.o',
     BUILD_DIR + '/lib/src/unknown_0D3020.o'
 ]
